// src/queries/model/categoryResponseDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * View Nest API
 * API documentation for View Nest application
 * OpenAPI spec version: 1.0
 */
import type { CategoryResponseDtoMdfcnDt } from './categoryResponseDtoMdfcnDt';
import type { CategoryResponseDtoMdfrId } from './categoryResponseDtoMdfrId';
import type { CategoryResponseDtoRgtrIp } from './categoryResponseDtoRgtrIp';

export interface CategoryResponseDto {
  /** 코드 */
  cd: string;
  /** 코드 분류명 */
  cdClsfNm: string;
  /** 코드 레벨 */
  cdLvl: number;
  /** 코드명 */
  cdNm: string;
  /** 코드 전체명 */
  cdWholNm: string;
  /** 카테고리 ID */
  id: number;
  /**
   * 수정일시
   * @nullable
   */
  mdfcnDt: CategoryResponseDtoMdfcnDt;
  /**
   * 수정자 ID
   * @nullable
   */
  mdfrId: CategoryResponseDtoMdfrId;
  /** 등록일시 */
  regDt: string;
  /** 등록자 ID */
  rgtrId: number;
  /**
   * 등록자 IP
   * @nullable
   */
  rgtrIp: CategoryResponseDtoRgtrIp;
}