// src/queries/model/productResponseDto.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * View Nest API
 * API documentation for View Nest application
 * OpenAPI spec version: 1.0
 */
import type { ProductResponseDtoCompany } from './productResponseDtoCompany';
import type { ProductResponseDtoPrdctSttsCd } from './productResponseDtoPrdctSttsCd';
import type { ProductResponseDtoRegister } from './productResponseDtoRegister';
import type { ProductResponseDtoUser } from './productResponseDtoUser';

export interface ProductResponseDto {
  /**
   * 추가 이미지1 경로
   * @nullable
   */
  addImgPath1Vl?: string | null;
  /**
   * 추가 이미지2 경로
   * @nullable
   */
  addImgPath2Vl?: string | null;
  /** 브랜드명 */
  brandNm: string;
  /**
   * 용기형태
   * @nullable
   */
  bttlShp?: string | null;
  /**
   * 구성
   * @nullable
   */
  cmpsVl?: string | null;
  /**
   * 브랜드사
   * @nullable
   */
  company?: ProductResponseDtoCompany;
  /**
   * 중량
   * @nullable
   */
  cpctVl?: string | null;
  /** 카테고리 */
  ctgrCd: string;
  /**
   * 삭제 여부
   * @nullable
   */
  delYn?: string | null;
  /**
   * 소개 자료 첨부
   * @nullable
   */
  explnAtchFilePathVl?: string | null;
  /**
   * 입수
   * @nullable
   */
  getVl?: string | null;
  /**
   * 입수 수량
   * @nullable
   */
  inRcvQntVl?: string | null;
  /**
   * 인허가인증파일
   * @nullable
   */
  lcpmtCertFilePathVl?: string | null;
  /**
   * 주요 효능
   * @nullable
   */
  mainEffcncVl?: string | null;
  /**
   * 주요 성분
   * @nullable
   */
  mainIgdVl?: string | null;
  /**
   * 수정 일시
   * @nullable
   */
  mdfcnDt?: string | null;
  /**
   * 수정자 ID
   * @nullable
   */
  mdfrId?: number | null;
  /**
   * 제조국
   * @nullable
   */
  mnftrNtnNm?: string | null;
  /**
   * 상품 금액
   * @nullable
   */
  prdctAmt?: string | null;
  /**
   * 상품 설명
   * @nullable
   */
  prdctCn: string | null;
  /** 상품정보아이디 */
  prdctId: number;
  /**
   * 상품 정보 내용
   * @nullable
   */
  prdctInfoCn?: string | null;
  /** 상품명 */
  prdctNm: string;
  /**
   * 상품 판매 채널
   * @nullable
   */
  prdctNtslChnVl?: string | null;
  /**
   * 제품 규격
   * @nullable
   */
  prdctSpcfctVl?: string | null;
  /**
   * 상품 상태 코드
   * @nullable
   */
  prdctSttsCd: ProductResponseDtoPrdctSttsCd;
  /**
   * 제안가
   * @nullable
   */
  prpslAmt?: string | null;
  /**
   * 제안서 페이지 번호
   * @nullable
   */
  prpslDocPageNo?: number | null;
  /**
   * 권유 고객
   * @nullable
   */
  rcmdtnCustNm?: string | null;
  /**
   * 권유 포인트
   * @nullable
   */
  rcmdtnPntCn?: string | null;
  /**
   * 수신 수량
   * @nullable
   */
  rcvQntVl?: string | null;
  /**
   * 등록 일시
   * @nullable
   */
  regDt?: string | null;
  /**
   * 등록자
   * @nullable
   */
  register?: ProductResponseDtoRegister;
  /**
   * 반려사유
   * @nullable
   */
  rjctRsnCn?: string | null;
  /**
   * 출고 일시
   * @nullable
   */
  rlsDt?: string | null;
  /** 유통기한 */
  rtlTermYmd: string;
  /**
   * 태그
   * @nullable
   */
  tagCd?: string | null;
  /**
   * 태그 목록 (쉼표로 구분된 문자열, 예: #태그1,#태그2) only for View, not for Update
   * @nullable
   */
  tags?: string | null;
  /**
   * 썸네일 누끼 이미지 경로
   * @nullable
   */
  thumbImgPath2Vl: string | null;
  /**
   * 썸네일 이미지 경로
   * @nullable
   */
  thumbImgPathVl: string | null;
  /**
   * 사용 방법 내용
   * @nullable
   */
  useMthdCn?: string | null;
  /**
   * 사용자 정보
   * @nullable
   */
  user?: ProductResponseDtoUser;
}